feat: overwrite costs through file (#1752)
* feat: introduce build_cost_data

* refactor: bring load_costs features into build_cost_data

* doc: improve documentation

* doc: add release note

* feat: enable scenario support through file parameter

* doc: add explicit description of the preparation stage

* fix: enable custom costs by default

* fix: fix units case

* doc: improve description of default values

* fix: resolve year wildcard ambiguity

* refactor: adjust naming convention

* replace year with planning_horizons

* remove expanding to all & overwriting only given costs without priotization

* update configtable accordingly

* Update rules/collect.smk


* year bugfix

* remove .format() in rules

* harmonize costs and tech_costs

* untangle params and config

* dtype change

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* typo

* Update config/config.default.yaml


* Update doc/configtables/costs.csv


* Update rules/build_electricity.smk


* rename build_cost_data to process_cost_data

* rename build_cost_data to process_cost_data ii

* clean snakemake inputs

* omit conig usage

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* calculate nyears from actual snapshots

* from manual to default nyears calculation

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Update doc/release_notes.rst


* omit year as a wildcard

* add 'all'-tech feature from Thomas

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* explicit mention of index_cols

* Apply suggestions from code review


---------
