+++ config/test/config.electricity.yaml
@@ -63,7 +63,6 @@ renewable:
     max_depth: false
     min_depth: false
 
-
 clustering:
   exclude_carriers: ["OCGT", "offwind-ac", "coal"]
   temporal:
