+++ doc/configtables/costs.csv
@@ -3,6 +3,7 @@ year,--,YYYY; e.g. '2030',Year for which to retrieve cost assumptions of ``resou
 version,--,vX.X.X or <user>/<repo>/vX.X.X; e.g. 'v0.5.0',Version of ``technology-data`` repository to use. If this string is of the form <user>/<repo>/<version> then costs are instead retrieved from ``github.com/<user>/<repo>`` at the <version> tag.
 social_discountrate,p.u.,float,Social discount rate to compare costs in different investment periods. 0.02 corresponds to a social discount rate of 2%.
 fill_values,--,float,Default values if not specified for a technology in ``resources/costs.csv``.
+custom_cost_fn,--,str or None,Path to the custom costs file. None if it should not be used. Default ``data/custom_costs.csv`` contains minor adjustments for stabilising the optimisation results.
 overwrites,--,Keys should be in the 'parameter' column of ``resources/costs.csv``. Values are dictionaries with keys from the  'technology' column of ``resources/costs.csv`` and float values.,"For the given parameters and technologies, assumptions about their parameter are overwritten the corresponding value of the technology."
 capital_cost,EUR/MW,Keys should be in the 'technology' column of ``resources/costs.csv``. Values can be any float.,"For the given technologies, assumptions about their capital investment costs are set to the corresponding value. Optional; overwrites cost assumptions from ``resources/costs.csv``."
 marginal_cost,EUR/MWh,Keys should be in the 'technology' column of ``resources/costs.csv``. Values can be any float.,"For the given technologies, assumptions about their marginal operating costs are set to the corresponding value. Optional; overwrites cost assumptions from ``resources/costs.csv``."
