+++ doc/release_notes.rst
@@ -9,6 +9,8 @@ Release Notes
 Upcoming Release
 ================
 
+* Feature: Introduce a new method to overwrite costs (https://github.com/PyPSA/pypsa-eur/pull/1752). Modifications to the default techno-economic assumptions can now be configured via `costs:custom_cost_fn`, which applies changes to the `resources/costs_{planning_horizons}.csv` files. The default configuration includes minor adjustments to stabilize optimization results. The existing implementation via `costs:overwrites` and `costs:capital_cost`/`costs:marginal_cost` parameters remains available but will be deprecated in a future release.
+
 * Fixed `AttributeError` in `prepare_sector_network.py` when running sector-coupled
   PyPSA-Eur with only one country and cluster. 
   (https://github.com/PyPSA/pypsa-eur/pull/1835)
