+++ rules/collect.smk
@@ -12,6 +12,19 @@ localrules:
     solve_sector_networks,
 
 
+rule process_costs:
+    input:
+        lambda w: (
+            [resources(f"costs_{config_provider('costs', 'year')(w)}_processed.csv")]
+            if config_provider("foresight")(w) == "overnight"
+            else expand(
+                resources("costs_{planning_horizons}_processed.csv"),
+                **config["scenario"],
+                run=config["run"]["name"],
+            )
+        ),
+
+
 rule cluster_networks:
     input:
         expand(
